⚡ Fix: 유저아이디 파라미터 없이도 접속 가능하게 수정, 카테고리 필요한 정보만 리턴

+++ src/main/java/com/trendithon/timetris/domain/mainpage/controller/CategoryController.java
@@ -5,10 +5,12 @@ import com.trendithon.timetris.domain.mainpage.dto.CategoryCreateDTO;
 import com.trendithon.timetris.domain.mainpage.dto.CategoryRequestDTO;
 import com.trendithon.timetris.domain.mainpage.dto.CategoryViewDTO;
 import com.trendithon.timetris.domain.mainpage.service.CategoryService;
+import com.trendithon.timetris.global.auth.jwt.TokenProvider;
 import com.trendithon.timetris.global.exception.ApiResponse;
 import com.trendithon.timetris.global.exception.CustomException;
 import com.trendithon.timetris.global.exception.enums.ErrorStatus;
 import com.trendithon.timetris.global.exception.enums.SuccessStatus;
+import jakarta.servlet.http.HttpServletRequest;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.core.Authentication;
 import org.springframework.web.bind.annotation.*;
@@ -21,42 +23,95 @@ import java.util.List;
 public class CategoryController {
 
     private final CategoryService categoryService;
+    private final TokenProvider tokenProvider;
 
-    @GetMapping("/{userId}")
-    public ApiResponse<List<Category>> readCategoryAll(Authentication authentication,
-                                                       @PathVariable long userId)
+    @GetMapping("")
+    public ApiResponse<List<CategoryViewDTO>> readCategoryAll(HttpServletRequest request)
     {
-        List<Category> categoryList = categoryService.readCategoryAll(userId);
+        String accessToken = tokenProvider.extractAccessToken(request).orElse(null);
+
+        Long userId = null;
+
+        if (accessToken != null) {
+            userId = tokenProvider.extractId(accessToken).orElse(null);
+        } else {
+            return ApiResponse.failure(ErrorStatus.NOT_LOGIN_ERROR);
+
+        }
+        if(userId == null) {
+            throw new CustomException(ErrorStatus.USER_NOT_FOUND_ERROR);
+        }
+
+        List<CategoryViewDTO> categoryList = categoryService.readCategoryAll(userId);
         return ApiResponse.success(SuccessStatus.OK, categoryList);
     }
 
     @PostMapping("")
-    public ApiResponse createCategory(Authentication authentication,
+    public ApiResponse<CategoryRequestDTO> createCategory(HttpServletRequest request,
                                       @RequestBody CategoryRequestDTO categoryRequestDTO)
     {
-        long userId = 1;
+        String accessToken = tokenProvider.extractAccessToken(request).orElse(null);
+
+        Long userId = null;
+
+        if (accessToken != null) {
+            userId = tokenProvider.extractId(accessToken).orElse(null);
+        } else {
+            return ApiResponse.failure(ErrorStatus.NOT_LOGIN_ERROR);
+
+        }
+        if(userId == null) {
+            throw new CustomException(ErrorStatus.USER_NOT_FOUND_ERROR);
+        }
+
         Category category = categoryService.createCategory(userId, categoryRequestDTO);
         if (category == null){
             throw new CustomException(ErrorStatus.INVALID_FORMAT_ERROR);
         }
-        return ApiResponse.success(SuccessStatus.OK);
+        return ApiResponse.success(SuccessStatus.OK, categoryRequestDTO);
     }
 
     @PutMapping("/{categoryId}")
-    public ApiResponse updateCategory(Authentication authentication,
+    public ApiResponse<CategoryViewDTO> updateCategory(HttpServletRequest request,
                                       @PathVariable long categoryId,
                                       @RequestBody CategoryViewDTO categoryViewDTO)
     {
-        long userId = 1;
+        String accessToken = tokenProvider.extractAccessToken(request).orElse(null);
+
+        Long userId = null;
+
+        if (accessToken != null) {
+            userId = tokenProvider.extractId(accessToken).orElse(null);
+        } else {
+            return ApiResponse.failure(ErrorStatus.NOT_LOGIN_ERROR);
+
+        }
+        if(userId == null) {
+            throw new CustomException(ErrorStatus.USER_NOT_FOUND_ERROR);
+        }
+
         categoryService.updateCategory(userId, categoryId, categoryViewDTO);
-        return ApiResponse.success(SuccessStatus.OK);
+        return ApiResponse.success(SuccessStatus.OK, categoryViewDTO);
     }
 
     @DeleteMapping("/{categoryId}")
-    public ApiResponse deleteCategory(Authentication authentication,
+    public ApiResponse deleteCategory(HttpServletRequest request,
                                       @PathVariable long categoryId)
     {
-        long userId = 1;
+        String accessToken = tokenProvider.extractAccessToken(request).orElse(null);
+
+        Long userId = null;
+
+        if (accessToken != null) {
+            userId = tokenProvider.extractId(accessToken).orElse(null);
+        } else {
+            return ApiResponse.failure(ErrorStatus.NOT_LOGIN_ERROR);
+
+        }
+        if(userId == null) {
+            throw new CustomException(ErrorStatus.USER_NOT_FOUND_ERROR);
+        }
+
         categoryService.deleteCategory(userId, categoryId);
         return ApiResponse.success(SuccessStatus.OK);
     }
